Added the Consumable api
Added the option to get all Appointmentgroup objects with a given status
Fixed the isAvailable Method, so it works properly/right
Refactored the isAvailable-Method to a service-class of its own. Now it can be used by the Employee and Resource services.
Added a datetime-format for the json-deserialization of dates
Refactored the RestExceptionHandler, so that it now sends the exception-message to the caller and does not show the exception-stack