Added a test for the appointments that might be affected by the ending of an availability, so a warning flag is set if the entity is not available anymore
Added functionality to the AvailabilityServiceImpl. It can now load all availabilities and save all changed availabilities
Refactored the creation of procedures/resources/employees, so that the availabilities will be saved first.
Refactored the update of procedures/resources/employees, so that all availabilities of the object will be loaded from the database first
Refactored Availability so that every availability is now saved in a separate document on the database